requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.34.27
botocore==1.34.27
celery==5.3.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.2.9
django-allauth==0.51.0
django-appconf==1.0.6
django-celery-beat==2.5.0
django-compressor==4.4
django-cors-headers==4.3.1
django-redis==5.4.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
kombu==5.3.5
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
twilio==8.12.0
typing-extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
